move more to abstract class
