+++ SingularityService/src/main/java/com/hubspot/singularity/hooks/NoOpLoadBalancerClient.java
@@ -8,38 +8,26 @@ import com.hubspot.singularity.SingularityDeploy;
 import com.hubspot.singularity.SingularityLoadBalancerUpdate;
 import com.hubspot.singularity.SingularityLoadBalancerUpdate.LoadBalancerMethod;
 import com.hubspot.singularity.SingularityRequest;
-import com.hubspot.singularity.SingularityTask;
+import com.hubspot.singularity.config.SingularityConfiguration;
+import com.hubspot.singularity.helpers.MesosProtosUtils;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-public class NoOpLoadBalancerClient implements LoadBalancerClient {
+public class NoOpLoadBalancerClient extends LoadBalancerClient {
 
   @Inject
-  public NoOpLoadBalancerClient() {}
+  public NoOpLoadBalancerClient(
+    SingularityConfiguration configuration,
+    MesosProtosUtils mesosProtosUtils
+  ) {
+    super(configuration, mesosProtosUtils);
+  }
 
   @Override
   public boolean isEnabled() {
     return false;
-  }
-
-  @Override
-  public SingularityLoadBalancerUpdate enqueue(
-    LoadBalancerRequestId loadBalancerRequestId,
-    SingularityRequest request,
-    SingularityDeploy deploy,
-    List<SingularityTask> add,
-    List<SingularityTask> remove
-  ) {
-    return new SingularityLoadBalancerUpdate(
-      LoadBalancerRequestState.SUCCESS,
-      loadBalancerRequestId,
-      Optional.empty(),
-      System.currentTimeMillis(),
-      LoadBalancerMethod.ENQUEUE,
-      Optional.empty()
-    );
   }
 
   @Override
@@ -89,15 +77,6 @@ public class NoOpLoadBalancerClient implements LoadBalancerClient {
 
   @Override
   public List<LoadBalancerUpstream> getUpstreamsForRequest(String singularityRequestId) {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public List<LoadBalancerUpstream> getUpstreamsForTasks(
-    List<SingularityTask> tasks,
-    String requestId,
-    Optional<String> loadBalancerUpstreamGroup
-  ) {
     return Collections.emptyList();
   }
