+++ SingularityService/src/test/java/com/hubspot/singularity/data/ValidatorTest.java
@@ -56,7 +56,8 @@ public class ValidatorTest extends SingularitySchedulerTestBase {
         priorityManager,
         disasterManager,
         agentManager,
-        uiConfiguration
+        uiConfiguration,
+        testingLbClient
       );
   }
